Added file operations and changed list to use 64 bit numbers. Beginnings of shaders.

+++ src/Coco/Core/Services/EngineServiceManager.cpp
@@ -6,7 +6,7 @@ namespace Coco
 	EngineServiceManager::~EngineServiceManager()
 	{
 		// Destroy services in reverse order that they were added
-		for (int i = _services.Count() - 1; i >= 0; i--)
+		for (unsigned long long i = _services.Count() - 1; i >= 0; i--)
 		{
 			_services[i].reset();
 		}
